requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
pymongo==4.10.1
google-auth==2.37.0
anthropic==0.40.0
python-dotenv==1.0.1
requests==2.32.5
